[refactor] small refactor of codebase, add some andpoints
