+++ resources/js/src/common/api/endpoints/__example__.js
