+++ resources/js/src/common/api/endpoints/grades.js
@@ -1,0 +1,7 @@
+/**
+ * Получить список классов
+ * @return { Object<{ uri: String }>}
+ **/
+export function getAll() {
+    return { uri: "grades" };
+}
