+++ resources/js/src/common/api/endpoints/subjects.js
@@ -1,0 +1,7 @@
+/**
+ * Получить список субьектов
+ * @return { Object<{ uri: String }>}
+ **/
+export function getAll() {
+    return { uri: "subjects" };
+}
