+++ resources/js/src/common/api/endpoints/tasks.js
@@ -1,0 +1,7 @@
+/**
+ * Получить список задач
+ * @return { Object<{ uri: String }>}
+ **/
+export function getAll() {
+    return { uri: "tasks" };
+}
