+++ resources/js/src/models/Subject.js
@@ -1,0 +1,6 @@
+import { Model } from "./Model";
+
+class Subject extends Model {
+    grade;
+    tasks = [];
+}
