+++ resources/js/src/models/Task.js
@@ -1,0 +1,5 @@
+import { Model } from "./Model";
+
+class Task extends Model {
+
+}
